✨ rename components "offers" > "listings"
